+++ models/__init__.py
@@ -2,3 +2,4 @@ from . import project_task
 from . import project_task_skill_group
 from . import project_project
 from . import project_capacity_report
+from . import report_capacity_report